Added in new Functions and Removed 2D array
The 2d array had been implemented to let me convert ASCII to Hexadecimal and back again. Turns out this was not needed, as you can use a typecast to cast a character as an integer. Therefore, the process was altered.
I also started the process to allow me to implement pthreading to the code. This will allow me to use Continuous RQ later on down the road.